Andre W. pointed out it should probably be qreal just in case that's different from float; works on SPARC and x86.
svn path=/trunk/KDE/kdebase/workspace/libs/plasma/; revision=747411

+++ layouts/layout.cpp
@@ -219,7 +219,7 @@ QSizeF Layout::minimumSize() const
 }
 QSizeF Layout::maximumSize() const
 {
-    return QSizeF(std::numeric_limits<float>::infinity(),std::numeric_limits<float>::infinity());
+    return QSizeF(std::numeric_limits<qreal>::infinity(),std::numeric_limits<qreal>::infinity());
 }
 
 void Layout::startAnimation()
